refactor the beryllium.rs into common for better reuse
