+++ bootloader/Cargo.toml
@@ -9,3 +9,4 @@ edition = "2021"
 bitflags = "2.3.1"
 uefi = { version="0.24.0", features = ["alloc", "global_allocator", "panic-on-logger-errors", "logger"] }
 uefi-services = "0.21.0"
+common = { path = "../common" }
