+++ common/src/lib.rs
@@ -1,4 +1,5 @@
 #![no_std]
 
+pub mod beryllium;
 pub mod font;
 pub mod framebuffer;
