+++ kernel/src/aarch64/mod.rs
@@ -1,3 +1,2 @@
 pub mod arch_api;
 mod asm;
-mod beryllium;
